requirements.txt
blinker==1.6.2
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.3
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.3.2
fonttools==4.39.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
protobuf==4.21.12
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tenacity==8.2.2
tzdata==2023.3
Werkzeug==2.3.4
wordcloud==1.9.1.1
